pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "foxtronpi-pyclient"
version = "0.1.0"
authors = [
  { name="Sirius Wu", email="Sirius.CC.Wu@foxtronev.com" },
]
description = "An example python project to control the FoxtronEV's car model D31X under project FoxtronPI with python-doipclient and python-udsoncan."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]
license = { file = "LICENSE.txt" }
dependencies = [
    "python-doipclient",
    "python-udsoncan",
]

[tool.foxtronpi-pyclient]
# Configuration for the FoxtronPI client application.
#
# SECURITY WARNING:
# It is critical to avoid storing secret keys directly in version-controlled files.
# This configuration section defines the names of the environment variables
# that the application should use to load the security keys at runtime.
security_key_env_vars = [
  "D31X_SECURITY_KEY_1",
  "D31X_SECURITY_KEY_2",
  "D31X_SECURITY_KEY_3",
]